enemies nd beautiful stuf

+++ Background.hpp
@@ -20,7 +20,7 @@ public:
     bool colision(sf::Vector2f pos);
     void readLevel(std::string lvlDesc);
     void draw(sf::RenderTarget *target);
-    bool circleColision(sf::Vector2i pos, float rad);
+    bool circleColision(sf::Vector2f pos, float rad);
     sf::Vector2i getIntersection(sf::Vector2i mousePos);
 
     bool colision(float x, float y);
